Test task cache
